Restructure API files into subfolder
