+++ pyproject.toml
@@ -10,6 +10,7 @@ requires-python = ">=3.13"
 dependencies = [
     "alembic>=1.16.4",
     "fastapi[standard]>=0.116.1",
+    "nicegui>=2.21.1",
     "psycopg2-binary>=2.9.10",
     "python-decouple>=3.8",
     "sqlmodel>=0.0.24",
@@ -29,7 +30,7 @@ cov = "pytest --cov=src --cov-report=term-missing"
 lint = "uv tool run ruff check ."
 format = "uv tool run ruff format ."
 tc = "uv tool run ty check ."
-apirun = "uv run fastapi dev src/main.py"
+apirun = "uv run fastapi dev src/api/main.py"
 
 [tool.poe.tasks.dummy]
 cmd = "uv run scripts/generate_dummy_data.py"
