+++ src/api/routers/categories.py
